+++ Includes/output.hpp
@@ -7,15 +7,15 @@
 
 using namespace std;
 
-string main_name(double H, double J, int size, double k, char shape, char hamil);
+string main_name(double H, double J, double size, double k, char shape, char hamil);
 
 string main_name_d(double H, double J, double amean, double asd, double k, char shape, char hamil);
 
-string full_name(double H, double J, int size, double k, char shape, char hamil, double T);
+string full_name(double H, double J, double size, double k, char shape, char hamil, double T);
 
 string full_name_d(double H, double J, double amean, double asd, double k, char shape, char hamil, double T);
 
-void create_folder(double H, double J, int size, double k, char shape, char hamil);
+void create_folder(double H, double J, double size, double k, char shape, char hamil);
 
 void create_folder_d(double H, double J, double amean, double asd, double k, char shape, char hamil);
 
@@ -30,7 +30,7 @@ void print_avs(string avname, vector<double>& allener, vector<double>& allmag,
     vector<double>& allsmagz, vector<int>& allnums, double H, char hamil,
     double g_lattsize, double g_slattsize, double T);
 
-void init_avs(string avname, double H, double J, double k, int size, double Tmin,
+void init_avs(string avname, double H, double J, double k, double size, double Tmin,
     double Tmax);
 
 #endif
